// src/types.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN = 'ADMIN',
  PENGURUS = 'PENGURUS',
  ANGGOTA = 'ANGGOTA',
}

export enum AttendanceMode {
  QR = 'QR',
  GPS = 'GPS',
  HYBRID = 'HYBRID' // GPS + QR
}

export enum AgendaCategory {
  RAPAT = 'RAPAT',
  PELATIHAN = 'PELATIHAN',
  SOSIAL = 'SOSIAL',
  DARURAT = 'DARURAT',
  LAINNYA = 'LAINNYA'
}

// Base Interface for Soft Delete & Integrity & Multi-Tenancy
export interface BaseEntity {
  deleted?: boolean;
  deletedAt?: string;
  deletedBy?: string;
  _checksum?: string;
  _tainted?: boolean;
  orgId?: string; // Multi-Organization Isolation
}

export interface Organization extends BaseEntity {
    id: string;
    name: string;
    pinCode: string; // For Secure Export
    settings: {
        slaTarget: number; // e.g. 85%
        allowMultiDevice: boolean;
        strictMode: boolean;
    };
}

export interface User extends BaseEntity {
  id: string;
  nia?: string; 
  name: string;
  division: string;
  role: Role;
  password?: string; // Only used for client-side legacy auth, otherwise managed by Supabase Auth
  phone: string;
  email?: string;
  isActive: boolean;
  avatar?: string;
  joinDate?: string;
  trustScore?: number;
  totalPoints?: number;
  photoReference?: string;
  badges?: string[]; 
  
  // Security
  deviceId?: string;
  lastLogin?: string;
  
  preferences?: {
    notifications: boolean;
    darkMode?: boolean;
    language?: string;
  };
  adminNotes?: { id: string, content: string, date: string, author: string }[];
}

export interface GeoLocation {
    id: string;
    name: string;
    lat: number;
    lng: number;
    radius: number;
    type: 'INDOOR' | 'OUTDOOR';
}

export interface Event extends BaseEntity {
  id: string;
  name: string;
  date: string; 
  startTime: string;
  endTime: string;
  locationName: string; 
  locations?: GeoLocation[]; 
  latitude: number; 
  longitude: number; 
  radiusMeters: number; 
  mode: AttendanceMode;
  status: 'BERLANGSUNG' | 'SELESAI' | 'DITUNDA' | 'BATAL' | 'AKAN_DATANG' | 'PENUH';
  locked?: boolean; 
  resumeDeadline?: string;
  resumedAt?: string;
  lateToleranceMinutes?: number;
  
  // Capacity
  maxCapacity?: number;
  currentAttendees?: number;
}

export interface AttendanceRecord extends BaseEntity {
  id: string;
  userId: string;
  eventId: string;
  timestamp: string;
  status: 'PRESENT' | 'LATE' | 'ABSENT' | 'PERMISSION' | 'ON_HOLD' | 'CANCELLED';
  latitude?: number;
  longitude?: number;
  trustScore: number; 
  trustBadges?: string[]; 
  deviceInfo?: string;
  notes?: string;
  proofPhoto?: string; 
  synced?: boolean; 
  
  // Enterprise Upload Data
  metadata?: {
    file_hash?: string;
    file_size?: number;
    storage_path?: string;
    upload_retry_count?: number;
  }
}